renames jsx files to js
