+++ webpack.production.config.js
@@ -4,7 +4,7 @@ const webpack = require('webpack');
 const path = require('path');
 
 module.exports = {
-    entry: "./src/js/app.jsx",
+    entry: "./src/js/app.js",
     output: {
         path: path.join(__dirname, 'dist'),
         filename: 'bundle.js'
